Add placeholderTextColor
This theme, especially the dark variant, did not handle text input
placeholders very well. For example, if you open the DuckDuckGo action
to do a web search, but had not typed any text into the input yet, you'd
barely be able to see the "DuckDuckGo" input placeholder text because it
was so dark against the LaunchBar background.

This just explicitly styles that placeholder text.